refactor: getMyPost post도메인으로 이동 및 api 레이어 분리
